Protected Sheets as separate nuget package
