+++ src/SheetToObjects.Adapters.GoogleSheets/SheetToObjects.Adapters.GoogleSheets.csproj
@@ -15,7 +15,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Google.Apis.Sheets.v4" Version="1.35.1.1338" />
     <PackageReference Include="Refit" Version="4.4.17" />
   </ItemGroup>
