+++ src/SheetToObjects.Adapters.ProtectedGoogleSheets/GoogleClientServiceFactory.cs
@@ -3,7 +3,7 @@ using Google.Apis.Auth.OAuth2;
 using Google.Apis.Services;
 using Google.Apis.Sheets.v4;
 
-namespace SheetToObjects.Adapters.GoogleSheets
+namespace SheetToObjects.Adapters.ProtectedGoogleSheets
 {
     public class GoogleClientServiceFactory : ICreateGoogleClientService
     {
