+++ src/SheetToObjects.Adapters.ProtectedGoogleSheets/ISheetsServiceWrapper.cs
@@ -1,7 +1,7 @@
 ﻿using System.Threading.Tasks;
 using Google.Apis.Sheets.v4.Data;
 
-namespace SheetToObjects.Adapters.GoogleSheets
+namespace SheetToObjects.Adapters.ProtectedGoogleSheets
 {
     public interface ISheetsServiceWrapper
     {
