--- src/SheetToObjects.Adapters.ProtectedGoogleSheets/Interfaces/IProvideSheet.cs
@@ -1,10 +1,0 @@
-﻿using System.Threading.Tasks;
-using SheetToObjects.Lib;
-
-namespace SheetToObjects.Adapters.ProtectedGoogleSheets.Interfaces
-{
-    public interface IProvideSheet
-    {
-        Task<Sheet> GetAsync(string jsonFile, string appName, string sheetId, string range);
-    }
-}
