+++ src/SheetToObjects.ConsoleApp/ExcelApp.cs
@@ -29,7 +29,6 @@ namespace SheetToObjects.ConsoleApp
                 return _sheetMapper.Map<SuperstoreModel>(sheet);
             });
 
-            Console.WriteLine("===============================================================");
             foreach (var error in result.Item1.ValidationErrors)
             {
                 Console.WriteLine($"Column: {error.ColumnName} | Row: {error.RowIndex} | Message: {error.ErrorMessage}");
