+++ src/SheetToObjects.ConsoleApp/Files/profiles-with-validation-errors.csv
