+++ src/SheetToObjects.ConsoleApp/SheetToObjects.ConsoleApp.csproj
@@ -16,6 +16,7 @@
     <ProjectReference Include="..\SheetToObjects.Adapters.Csv\SheetToObjects.Adapters.Csv.csproj" />
     <ProjectReference Include="..\SheetToObjects.Adapters.GoogleSheets\SheetToObjects.Adapters.GoogleSheets.csproj" />
     <ProjectReference Include="..\SheetToObjects.Adapters.MicrosoftExcel\SheetToObjects.Adapters.MicrosoftExcel.csproj" />
+    <ProjectReference Include="..\SheetToObjects.Adapters.ProtectedGoogleSheets\SheetToObjects.Adapters.ProtectedGoogleSheets.csproj" />
     <ProjectReference Include="..\SheetToObjects.Lib\SheetToObjects.Lib.csproj" />
   </ItemGroup>
 
@@ -26,7 +27,10 @@
     <None Update="app-settings.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
-    <None Update="Files\profiles.csv">
+    <None Update="Files\profiles-without-validation-errors.csv">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="Files\profiles-with-validation-errors.csv">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
     <None Update="Files\profiles.xlsx">
