+++ src/SheetToObjects.Specs/Adapters/ProtectedGoogleSheetToSheetConverterSpecs.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using FluentAssertions;
 using Google.Apis.Sheets.v4.Data;
 using SheetToObjects.Adapters.GoogleSheets;
+using SheetToObjects.Adapters.ProtectedGoogleSheets;
 using SheetToObjects.Specs.Builders;
 using Xunit;
