+++ src/SheetToObjects.Specs/SheetToObjects.Specs.csproj
@@ -12,13 +12,14 @@
     <PackageReference Include="xunit" Version="2.3.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
     <DotNetCliToolReference Include="dotnet-xunit" Version="2.3.1" />
-    <PackageReference Include="Google.Apis.Sheets.v4" Version="1.35.1.1338" />
+    <PackageReference Include="Google.Apis.Sheets.v4" Version="1.36.1.1387" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\SheetToObjects.Adapters.Csv\SheetToObjects.Adapters.Csv.csproj" />
     <ProjectReference Include="..\SheetToObjects.Adapters.GoogleSheets\SheetToObjects.Adapters.GoogleSheets.csproj" />
     <ProjectReference Include="..\SheetToObjects.Adapters.MicrosoftExcel\SheetToObjects.Adapters.MicrosoftExcel.csproj" />
+    <ProjectReference Include="..\SheetToObjects.Adapters.ProtectedGoogleSheets\SheetToObjects.Adapters.ProtectedGoogleSheets.csproj" />
     <ProjectReference Include="..\SheetToObjects.Lib\SheetToObjects.Lib.csproj" />
   </ItemGroup>
