+++ src/SheetToObjects.sln
@@ -22,6 +22,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		Common.props = Common.props
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SheetToObjects.Adapters.ProtectedGoogleSheets", "SheetToObjects.Adapters.ProtectedGoogleSheets\SheetToObjects.Adapters.ProtectedGoogleSheets.csproj", "{C6EA22A6-873B-44BC-9C4E-AD01227FBDA0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -56,6 +58,10 @@ Global
 		{BB23F27B-8F37-4E22-89BE-C61268C1F75A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BB23F27B-8F37-4E22-89BE-C61268C1F75A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BB23F27B-8F37-4E22-89BE-C61268C1F75A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C6EA22A6-873B-44BC-9C4E-AD01227FBDA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C6EA22A6-873B-44BC-9C4E-AD01227FBDA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C6EA22A6-873B-44BC-9C4E-AD01227FBDA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C6EA22A6-873B-44BC-9C4E-AD01227FBDA0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
